✅ Rewrite tests dealing with `scheduler`
